requirements.txt
PyNaCl
discord.py
Pillow
aiohttp
docarray
jina
shortid
